src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-root-js-api-expected.txt

CONSOLE ERROR: line 21: Uncaught TypeError: shadowHost.createShadowRoot is not a function
Tests for ShadowRoot JS APIs. Can only run within DRT

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.ShadowRoot is defined.
PASS new window.ShadowRoot(document.createElement('div')) threw exception TypeError: Illegal constructor.
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
